pyproject.toml
[project]
name = "chancla-mix-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=3.3.0",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "spotipy>=2.25.2",
]
